Renaming note files to single form
