+++ src/database/entities/note.entity.ts
