+++ src/database/entities/user.entity.ts
@@ -1,6 +1,6 @@
 import { UserRole } from '../../types/enums';
 import { Column, Entity, OneToMany, PrimaryGeneratedColumn } from 'typeorm';
-import { NoteEntity } from './notes.entity';
+import { NoteEntity } from './note.entity';
 
 @Entity({ name: 'users' })
 export class UserEntity {
