+++ src/resource/notes/dto/note.dto.ts
