+++ src/resource/notes/note.module.ts
@@ -1,8 +1,8 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { NoteEntity } from 'src/database/entities/notes.entity';
-import { NoteService } from './notes.service';
-import { NoteController } from './notes.controller';
+import { NoteEntity } from 'src/database/entities/note.entity';
+import { NoteService } from './note.service';
+import { NoteController } from './note.controller';
 import { UserEntity } from 'src/database/entities/user.entity';
 
 @Module({
